entities/pyproject.toml
[project]
name = "user-registration"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[all] (>=0.115.11,<0.116.0)",
    "alembic (>=1.15.1,<2.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "pycryptodome (>=3.21.0,<4.0.0)",
    "python-jose[cryptography] (>=3.4.0,<4.0.0)",
    "aiosmtplib (>=4.0.0,<5.0.0)",
    "pytest (>=8.3.5,<9.0.0)",
    "pytest-asyncio (>=0.26.0,<0.27.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "aiosqlite (>=0.21.0,<0.22.0)",
    "phonenumbers (>=9.0.2,<10.0.0)",
    "selenium (>=4.31.0,<5.0.0)",
    "undetected-chromedriver (>=3.5.5,<4.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "celery[asyncio] (>=5.5.3,<6.0.0)",
    "redis (>=5.2.1,<6.0.0)",
    "boto3 (>=1.38.8,<2.0.0)",
    "asgiref (>=3.8.1,<4.0.0)",
    "nest-asyncio (>=1.6.0,<2.0.0)",
    "ruff (>=0.12.4,<0.13.0)",
]

[tool.ruff]
line-length = 119
target-version = "py312"
select = ["I"]
fixable = ["I"]
unfixable = ["F401", "F841"]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
